requirements.txt
Flask
Werkzeug
chess
Pillow
berserk
requests